backend/requirements.txt
fastapi==0.115.8
uvicorn==0.27.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
openai==1.61.1
gunicorn==20.1.0
python-dotenv==0.19.0
Werkzeug==3.1.3
flask-login==0.6.3
psycopg2-binary==2.9.10
SQLAlchemy==2.0.38
